requirements.txt
streamlit
groq
python-dotenv
pandas
reportlab
matplotlib
numpy
